MPT-16437 Reorganise resource billing mixins structure
